Fixed Karma System.
Added User Authentication to All Module Endpoints
Added Authentication to get Active Modules

userFunctionAuthorisations is a simple function to confirm user is Authorised to use a function. (This function does not test if a user Has access to a module).

bearer_decode is a function to decode a JWT token and return to user if Successful